fixed issues with newly flashed devices

+++ src/example_config.hpp
@@ -7,7 +7,7 @@
 #define LED_PIN 5
 #define BLINK_BRIGHTNESS 0.7
 
-#define LED_COUNT_FOR_ONE_SECOND (int)(NUM_LEDS / 60)
+#define LED_COUNT_FOR_ONE_SECOND (int)((double)NUM_LEDS / 60.0)
 #define DIMM_VALUE 0xFF
 #define DIMM_ADD_VALUE 0xAA
